+++ src/client/app/components/maps/MapsDetailComponent.tsx
@@ -12,7 +12,7 @@ import { Link } from 'react-router-dom';
 import TooltipHelpContainer from '../../containers/TooltipHelpContainer';
 import TooltipMarkerComponent from '../TooltipMarkerComponent';
 import { removeUnsavedChanges } from '../../actions/unsavedWarning';
-import store from '../../index';
+import { store }  from '../../store';
 import UnsavedWarningContainer from '../../containers/UnsavedWarningContainer';
 import HeaderComponent from '../../components/HeaderComponent';
