+++ src/client/app/components/unit/EditUnitModalComponent.tsx
@@ -2,7 +2,7 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 import * as React from 'react';
-import store from '../../index';
+import { store }  from '../../store';
 //Realize that * is already imported from react
 import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
